Implemented a DataWatch object for every node found either with self._get_node_from_provider, or self._get_nodes_from_provider.
In the event that the DataWatch object is created
by self._get_nodes_from_provider, the callbacks
are executed by self._get_nodes_from_provider. Any
other time the DataWatch object detects a change
though, it triggers the callbacks that are registered
to watch the path of that DataWatch object.