Fix 'fork' handling (issue #6):
When we detected a fork, we were trying to just stop and restart
our ZK connection. This is dangerous and unlikely to work.

Instead, we give up on our broken Kazoo client object and we
recreate it completely by walking through our _connect() method
again. As soon as we've called that method, we walk through
all of our Watcher and Registration objects and rebuild them
completely.

This effectively orphans our old Watcher/Registration and KazooClient
objects and lets them get garbage collected.