Do not try to close the original ZK connection in a fork.
In the event of a fork, do not bother trying to close the
original KazooClient object because we don't have access
to its threads. The stop will fail, and just delays the code
a few seonds while it tries.

Just create a new KazooClient object and move on with life.