Closes #9: Catch kazoo.exceptions.NoNodeError in _update_root_data()
When the _update_root_data() callback is executed, its passed in 'stat'
data that indicates whether or not the Node exists in Zookeeper or not.

Because nd_service_registry.shims.ZookeeeperClient creates an
artificial delay between Kazoo calls, its possible that the backend Node
path may become deregistered (expires) before zk.get() can be called. If
this happens, a NoNodeError exception is thrown and the app breaks.

Rather than using an 'if stat' statement, simply use a try/except block
and always request the node data. If its not there, handle it properly
and move on.

Change-Id: I3afd14c8cb321854925100052a7735b7420c2286