Closes #10: Implements custom throttling rules for ZK API calls
The ZookeeperClient API throttling code is moved entirely into the
shims.ZookeeperClient class. This allows us to leverage class-level
variables that define the number of calls to throttle and the average
time between them.

We now expose a set_rate_limiter() method in the ZookeeperClient that
the KazooServiceRegistry class uses to define the rate limit settings.
These settings are considered global for the entire ZookeeperClient
object now as well -- gets, sets, etc are all included.

Setting the rate_limit settings to None, False or 0 will disable them.

Change-Id: I783cffa56bcf35e3d799ccaf63eb32c1b17ab19b