Issue #16: Implement a 'wait' timer on the Lock object.
This patch leverages the latest Kazoo code to implement a 'wait' timer
on the Lock object. This allows a user to ask for a lock and wait up to
XX seconds for it before bailing out and performing some other
operation.

We also now set the default wait time to 300 seconds, preventing a full
deadlock of your application in the event that a Lock never becomes
available.

Change-Id: Ia13c5b4ba46ec05a7c3080f6696220afb3905b40