Add get_state() method to KazooServiceRegistry.
The get_state() method returns the current known
conection state of the KazooServceRegistry object to
Zookeeper. This state is managed by the _state_listener()
method and is updated any time a connection state change
is detected.

get_state() also accepts an optional callback argument
that will add a callback to the state listener and update
that callback any time the connection state changes.