Fix DummyWatcher creation behavior when no ZK connection.
When the @_health_check decorator fails, we try to create
a DummyWatcher from the cache and return that value. When it fails,
we were simply returning false and dying off.

This patch changes the behavior to still return False, but also
creates a blank DummyWatcher in the background. This DummyWatcher
object will later be converted into a real Watcher object when
the Zookeeper connection is re-established, and the appropriate
Callbacks will be executed.

This new behavior allows a client to request data, handle the failure
howevery they wish to, but then recover from the failure dynamically
as soon as the connection comes back.