Issue #38: Monitor non-existent paths for children changes.
This patch deals with issue #38. If you create a Watcher() on a path
that doesn't exist yet, Kazoo cannot create a ChildrenWatch object on
that path. In the past, we logged this event and moved on ... which
meant that when the path was created later on (and children were added),
your app would not be notified.

Now we register a second DataWatch on that non-existent path and look
for the state of the path to change. When it changes, we check if a
previous ChildrenWatch had been created. If not, we create a new one and
then record that. If one had already been created, we quickly exit the
function.

I've provided full integration tests for this behavior, but I have not
yet created standard Unit tests. I will do unit tests in a separate
checkin.