Issue 49: Don't recursively set ACL permissions on created paths.
This commit changes the ACLs used when recursively creating
paths in Zookeeper. If a path is nested (/foo/bar/baz/host:22),
we want /foo/bar/baz/host:22 and /foo/bar/baz to be created with
the ACL configured when the KazooServiceRegistry class was
instantiated. However, /foo/bar we want to have the default OPEN
ACL permissions so that other clients can write into the path
as necessary.