Add Kazoo 2.0 support as well as tests.
Kazoo 2.0 modified the KazooClient.add_auth() method to actually receive
and get the data for the asynchronous add authentication call. This
deadlocks when its wrapped inside of a KazooClient.retry() method.

The fix is to call add_auth_async() directly and ignore the result.
Additionally, I've thrown in tests that triggered this behavior, and
added Kazoo 2.0 to the testing matrix for Travis.