Handle TypeErrors when the Kazoo recipe passes back the wrong data through a callback.
Python-zk issue #255 is opened for this -- but it seems that we
occasionally see a Boolean returned to our callback function rather than
a list, which broke the code. Although this isn't supposed to happen, we
can handle the issue gracefully with a warning.

Also, handle Kazoo 2+'s Ping logger that moved to a new logger.