Support kazoo up to version 2.5.0.
Support newer versions of kazoo that seems to work with python 3.7.
This passes the unit tests and integration tests.

I had trouble getting 2.6.0+ working and it looks like it's b/c it
requires a new version of zookeeper for testing. Simply upgrading that
version wasn't enough, so this requires more investigation.